+++ utils/chart.types.ts
@@ -1,0 +1,6 @@
+export interface ChartsType {
+  time: string;
+  id: string;
+  value_area: string;
+  value_bar: string;
+}
